[rework]: archi

+++ src/instructions/Sub.cpp
@@ -11,19 +11,19 @@
 static void createInputState(QRegister &reg, size_t n)
 {
     reg.h(n);
-    for (int i = 0; i < n; i++)
+    for (size_t i = 0; i < n; i++)
         reg.cp(pow(2, i + 1), n, reg, n - (i + 1));
 }
 
 static void evolveQFTState(QRegister &reg_a, QRegister &reg_b, size_t n)
 {
-    for (int i = 0; i <= n; i++)
+    for (size_t i = 0; i <= n; i++)
         reg_a.cp(pow(2, i), n , reg_b, n - i);
 }
 
 static void inverseQFT(QRegister &reg, size_t n)
 {
-    for (int i = 0; i < n; i++)
+    for (size_t i = 0; i < n; i++)
         reg.cp(-1 * pow(2, n - i), n, reg, i);
     reg.h(n);
 }
@@ -34,11 +34,11 @@ static void doSub(QRegister &q1, QRegister &q2)
     q1.setValue(q2.getValue() - q1.getValue());
 
     q1.x();
-    for (int i = 0; i <= size; i++)
+    for (size_t i = 0; i <= size; i++)
         createInputState(q1, size - i);
-    for (int i = 0; i <= size; i++)
+    for (size_t i = 0; i <= size; i++)
         evolveQFTState(q1, q2, size - i);
-    for (int i = 0; i <= size; i++)
+    for (size_t i = 0; i <= size; i++)
         inverseQFT(q1, i);
     q1.x();
 }
